+++ src/sha1.h
@@ -21,7 +21,7 @@ typedef struct Sha1State
 
 void sha1Init(Sha1State* state);
 void sha1Count(Sha1State* state, const void* data);
-void sha1Tail(Sha1State* state,void* data, uint8_t currentBytes, uint64_t totalBytes);
+void sha1Tail(Sha1State* state,void* data, size_t currentBytes, uint64_t totalBytes);
 void sha1Result(Sha1State* state, char* result);
 
 void sha1(const void* data, size_t length, char* result);
